scene-app: Refactor to use SceneAppPage for demos
