--- packages/scenes-app/src/components/SecretInput/index.tsx
@@ -1,1 +1,0 @@
-export * from './SecretInput';
