+++ packages/scenes-app/src/constants.ts
@@ -3,7 +3,7 @@ import pluginJson from './plugin.json';
 export const PLUGIN_BASE_URL = `/a/${pluginJson.id}`;
 
 export enum ROUTES {
-  Home = 'home',
+  Home = '',
   Demos = 'demos',
   WithTabs = 'page-with-tabs',
   WithDrilldown = 'page-with-drilldown',
