+++ packages/scenes-app/src/demos/gridWithRow.tsx
@@ -1,0 +1,115 @@
+import {
+  VizPanel,
+  SceneGridLayout,
+  SceneGridRow,
+  SceneTimePicker,
+  SceneTimeRange,
+  EmbeddedScene,
+  SceneGridItem,
+  SceneAppPage,
+} from '@grafana/scenes';
+import { demoUrl } from '../utils/utils.routing';
+import { getQueryRunnerWithRandomWalkQuery } from './utils';
+
+export function getGridWithRowLayoutTest(): SceneAppPage {
+  return new SceneAppPage({
+    title: 'Grid with rows',
+    subTitle: 'SceneGridLayout demo with collapsible rows',
+    url: `${demoUrl('grid-layout-with-rows')}`,
+    getScene: () => {
+      return new EmbeddedScene({
+        body: new SceneGridLayout({
+          children: [
+            new SceneGridRow({
+              title: 'Row A',
+              key: 'Row A',
+              isCollapsed: true,
+              isDraggable: true,
+              isResizable: false,
+              y: 0,
+              x: 0,
+              children: [
+                new SceneGridItem({
+                  x: 0,
+                  y: 1,
+                  width: 12,
+                  height: 5,
+                  isResizable: true,
+                  isDraggable: true,
+                  body: new VizPanel({
+                    pluginId: 'timeseries',
+                    title: 'Row A Child1',
+                    key: 'Row A Child1',
+                  }),
+                }),
+                new SceneGridItem({
+                  x: 0,
+                  y: 5,
+                  width: 6,
+                  height: 5,
+                  isResizable: true,
+                  isDraggable: true,
+                  body: new VizPanel({
+                    pluginId: 'timeseries',
+                    title: 'Row A Child2',
+                    key: 'Row A Child2',
+                  }),
+                }),
+              ],
+            }),
+            new SceneGridRow({
+              title: 'Row B',
+              key: 'Row B',
+              isCollapsed: true,
+              y: 1,
+              children: [
+                new SceneGridItem({
+                  x: 0,
+                  y: 2,
+                  width: 12,
+                  height: 5,
+                  isResizable: false,
+                  isDraggable: true,
+                  body: new VizPanel({
+                    pluginId: 'timeseries',
+                    title: 'Row B Child1',
+                    key: 'Row B Child1',
+                  }),
+                }),
+                new SceneGridItem({
+                  x: 0,
+                  y: 7,
+                  width: 6,
+                  height: 5,
+                  isResizable: false,
+                  isDraggable: true,
+                  body: new VizPanel({
+                    pluginId: 'timeseries',
+                    title: 'Row B Child2',
+                    key: 'Row B Child2',
+                  }),
+                }),
+              ],
+            }),
+            new SceneGridItem({
+              x: 2,
+              y: 12,
+              width: 12,
+              height: 10,
+              isResizable: true,
+              isDraggable: true,
+              body: new VizPanel({
+                pluginId: 'timeseries',
+                title: 'Outsider',
+                key: 'Outsider',
+              }),
+            }),
+          ],
+        }),
+        $timeRange: new SceneTimeRange(),
+        $data: getQueryRunnerWithRandomWalkQuery(),
+        controls: [new SceneTimePicker({})],
+      });
+    },
+  });
+}
