--- packages/scenes-app/src/pages/Demos/index.ts
@@ -1,1 +1,0 @@
-export * from './Demos';
