--- packages/scenes-app/src/pages/Demos/scenes/nestedScene.tsx
@@ -1,58 +1,0 @@
-import {
-  VizPanel,
-  NestedScene,
-  SceneTimePicker,
-  SceneFlexLayout,
-  SceneTimeRange,
-  EmbeddedScene,
-  SceneFlexItem,
-} from '@grafana/scenes';
-import { getQueryRunnerWithRandomWalkQuery } from '../utils';
-
-export function getNestedScene(): EmbeddedScene {
-  return new EmbeddedScene({
-    body: new SceneFlexLayout({
-      direction: 'column',
-      children: [
-        new SceneFlexItem({
-          body: new VizPanel({
-            key: '3',
-            pluginId: 'timeseries',
-            title: 'Panel 3',
-          }),
-        }),
-        new SceneFlexItem({
-          body: getInnerScene('Inner scene'),
-        }),
-      ],
-    }),
-    $timeRange: new SceneTimeRange(),
-    $data: getQueryRunnerWithRandomWalkQuery(),
-    controls: [new SceneTimePicker({})],
-  });
-}
-
-export function getInnerScene(title: string) {
-  const scene = new NestedScene({
-    title: title,
-    canRemove: true,
-    canCollapse: true,
-    body: new SceneFlexLayout({
-      direction: 'row',
-      children: [
-        new SceneFlexItem({
-          body: new VizPanel({
-            key: '3',
-            pluginId: 'timeseries',
-            title: 'Data',
-          }),
-        }),
-      ],
-    }),
-    $timeRange: new SceneTimeRange(),
-    $data: getQueryRunnerWithRandomWalkQuery(),
-    actions: [new SceneTimePicker({})],
-  });
-
-  return scene;
-}
