--- packages/scenes-app/src/pages/Home/index.tsx
@@ -1,1 +1,0 @@
-export * from './Home';
