--- packages/scenes-app/src/pages/WithDrilldown/index.tsx
@@ -1,1 +1,0 @@
-export * from './WithDrilldown';
