--- packages/scenes-app/src/pages/WithTabs/index.tsx
@@ -1,1 +1,0 @@
-export * from './WithTabs';
